requirements.txt
boto3==1.9.122
botocore==1.12.122
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-cors-headers==2.5.2
django-extensions==2.1.6
django-filter==2.1.0
django-rest-auth==0.9.3
django-rest-multiple-models==2.1.3
django-storages==1.7.1
djangorestframework==3.9.2
docutils==0.14
idna==2.8
jmespath==0.9.4
Markdown==3.1
nltk==3.4.1
numpy==1.16.2
oauthlib==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
